Delete nd buffer
