+++ src/slangpy_ext/device/cursor_utils.h
@@ -14,7 +14,7 @@
 
 #include "utils/slangpy.h"
 #include "utils/slangpypackedarg.h"
-#include "utils/slangpystridedbufferview.h"
+#include "utils/slangpytensor.h"
 #include "sgl/device/buffer_cursor.h"
 
 #include <slang.h>
@@ -680,10 +680,10 @@ private:
         slang::TypeLayoutReflection* type_layout = self.slang_type_layout();
         auto kind = (TypeReflection::Kind)type_layout->getKind();
 
-        // Read uniforms for StridedBufferView unless it is being written directly to a pointer.
-        if (kind != TypeReflection::Kind::pointer && nb::isinstance<sgl::slangpy::StridedBufferView>(nbval)) {
-            auto view = nb::cast<sgl::slangpy::StridedBufferView*>(nbval);
-            nbval = view->uniforms();
+        // Read uniforms for NativeTensor unless it is being written directly to a pointer.
+        if (kind != TypeReflection::Kind::pointer && nb::isinstance<sgl::slangpy::NativeTensor>(nbval)) {
+            auto tensor = nb::cast<sgl::slangpy::NativeTensor*>(nbval);
+            nbval = tensor->uniforms();
         }
 
         switch (kind) {
@@ -746,11 +746,11 @@ private:
                 return;
             }
 
-            sgl::slangpy::StridedBufferView* sbview;
-            if (nb::try_cast<sgl::slangpy::StridedBufferView*>(nbval, sbview)) {
-                // If we have a StridedBufferView, write address of storage plus its byte offset.
-                uint64_t offset = static_cast<uint64_t>(sbview->offset()) * sbview->desc().element_layout->stride();
-                self.set_pointer(sbview->storage()->device_address() + offset);
+            sgl::slangpy::NativeTensor* tensor;
+            if (nb::try_cast<sgl::slangpy::NativeTensor*>(nbval, tensor)) {
+                // If we have a NativeTensor, write address of storage plus its byte offset.
+                uint64_t offset = static_cast<uint64_t>(tensor->offset()) * tensor->element_stride();
+                self.set_pointer(tensor->storage()->device_address() + offset);
                 return;
             }
