+++ src/slangpy_ext/slangpy_ext.cpp
@@ -65,8 +65,6 @@ SGL_PY_DECLARE(native_refl);
 SGL_PY_DECLARE(utils_crashpad);
 SGL_PY_DECLARE(utils_renderdoc);
 SGL_PY_DECLARE(utils_slangpy);
-SGL_PY_DECLARE(utils_slangpy_strided_buffer_view);
-SGL_PY_DECLARE(utils_slangpy_buffer);
 SGL_PY_DECLARE(utils_slangpy_function);
 SGL_PY_DECLARE(utils_slangpy_packedarg);
 SGL_PY_DECLARE(utils_slangpy_resources);
@@ -161,8 +159,6 @@ NB_MODULE(slangpy_ext, m_)
     SGL_PY_IMPORT(native_refl);
     SGL_PY_IMPORT(native_func);
     SGL_PY_IMPORT(utils_slangpy);
-    SGL_PY_IMPORT(utils_slangpy_strided_buffer_view);
-    SGL_PY_IMPORT(utils_slangpy_buffer);
     SGL_PY_IMPORT(utils_slangpy_function);
     SGL_PY_IMPORT(utils_slangpy_packedarg);
     SGL_PY_IMPORT(utils_slangpy_resources);
